Set up web3 with node url argument (gnosis/gp-v2-services#123)
We need a Web3 instance to make calls to the contract.
Next step is to set up the contracts instances with this web3.

### Test Plan
`cargo run --bin solver -- --node-url "http://localhost:8080"` to see that web3 is created. Doesn't actually need to have a node running at that address because no call is made.