Approve in Uniswap Interaction (gnosis/gp-v2-services#141)
This PR adds functionality to encode the approval tx of the sell token to Uniswap (which is required for the e2e test). In order to keep encoding synchronous, this is atm just passed in from above (in the future we should probably keep track of which approvals we have already given and only add them for newly traded tokens).

### Test Plan
Fully working e2e test (in two PRs). For now CI