Cache rust files in CI
I tested manually caching, using sccache and finally this pre made
action. The action works very well and is a clean one line solution.
Internally it caches the cargo registry and the target folder but not
crates from this workspace. The cache gets invalidated if Cargo.lock
changes.
sccache makes the build more complicated and doesn't give a speedup when
we already cache like this.