Verify EIP712 constants on deployed instance (gnosis/gp-v2-services#165)
This PR makes it so we cannot start the API if its internal magic numbers (in accordance with EIP 712) used for signature verification don't match the ones on the smart contract.

This is approximated by requiring that the hex representation of our local constants are contained in the deployed bytecode on chain. Since the identifiers are 32bytes, I think the chance for false negatives is negligible.

While the domain separator could be queried via a view method invocation, the OrderTypeHash cannot as it is stored as an internal constant on a linked library. Even if made public, we would have to explicitly expose it on the Settlement contract via a handwritten accessor or a StorageReading extension (cf. [this SO question](https://ethereum.stackexchange.com/questions/16082/solidity-accessing-a-library-constant-in-a-contract-that-imports-the-library)). It would also add up to one RPC roundtrip call per constant we want to check whereas fetching the bytecode and checking containment only requires only one roundtrip total.

### Test Plan
Change one byte in TypeHash constant, see we panic at startup