Use multi order solver (gnosis/gp-v2-services#170)
This PR makes use of the slightly less naive solver. For this we need to fetch the current "reserves" of the direct AMM between the tokens we are trying to match and pass it down into the logic.

To get the reserves, we need to query the `UniswapPair` (I added this contract artefact) from the UniswapFactory instance and then call `get_reserves` on it.

Overall the architecture is a little bit hacky as a lot of components are passed through large parts of the stack. I'd like to abstract "smart contract orders" such as Uniswap or Sushiswap pools (but later also Balancer, Curve, Prediction Market outcome tokens, etc) away from the solver interface so that we only hand an abstract set of orders which can be UserOrders or some form of smart contract order trait. Each solver can then work with the subset of orders they can handle.

However, in order to make sure the less naive solver works and to make the demo slightly more usable for external people it might be worthwhile merging like this and then refactor.

### Test Plan
I needed to remove the naive solver unit tests as we now rely on a smart contract call which is not yet easily mockable. However, the logic should still be covered by the multi order solver specific unit tests and the e2e test.

I can now trade e.g. STAKE for WXDAI on our xdai Demo interface (https://pr91--gpv2swap.review.gnosisdev.com/#/swap) directly without requiring a counter trade. This makes the demo more usable for external people who want to try it out and get a feeling for the e2e flow (in order to demonstrate coincedence of wants better we might have to prolong the solver sleep interval a bit to have more opportunity for overlap).

https://blockscout.com/poa/xdai/tx/0x1aedffc0c9a51bc1ace77303a8e5b622f4624a506a55d4ab9327abe15d3452a8/token-transfers is the first settlement using this solver.