Include executed amounts in order meta data
The solver needs at least the filled amount (depending on order type)
and I figured this could be useful information in general so we might as
well put them all in the api.
On the other hand we could decide that this shouldn't be part of the
public api because we do not want to guarantee that we provide it and
find another way to get this information to the solver.
The type is BigUint and not U256 because every amount except the filled
amount can technically sum up to an amount larger than U256::MAX over
multiple trades. The filled amount cannot because it is stored as a U256
in the smart contract.