Use balance fetcher (gnosis/gp-v2-services#246)
Part of #40 using the component created in #243.

We register tokens whenever an order is added and check balances when fetching orders. There, we also need to potentially create subscriptions (in case the order is coming from disk after a cold start).

If the cache is becoming too big we could potentially create a cleanup job/LRU strategy or the like to reduce the number of active subscriptions.

In order to pass the balance information to the solver and frontend, I created a new field on metadata. In the next PR I'm planning to add a filter to allow only fetching orders with valid balances.

### Test Plan
Added unit tests.