Filter insufficient balance (gnosis/gp-v2-services#250)
Closes #40 (for now)

Adding a new boolean to OrderFilter and enabling it for the default order requests, making sure we only return orders that are valid.

### Test Plan
Added a unit test. Also tested on Rinkeby with frontend pointing to localhost that when i create an order with max balance and then do a small transfer from that balance the order disappears from the orderbook.