Fix check that order has been settled
The order book api is supposed to still return settled orders when
queried by uid but the e2e test was using this condition. This works
with the in memory order book because it cleans up the orders when they
are settled where as the postgres order book does not.