Fetch Uniswap liquidity via batched call (gnosis/gp-v2-services#259)
This PR refactors the way we currently fetch Uniswap liquidity sources from using individual requests to a single RPC batch call. This is more efficient and will be needed if we want to include intermittent pools via the native token (GNO<>OWL should also add GNO<>ETH, ETH<>OWL)

On the way I noticed that the pool struct is no longer really necessary (it's kind of 1:1 replaced by the AMM liquidity subtype) so I removed it to simplify the logic in `solver/src/liquidity/uniswap.rs`. I also changed the way we compute the Uniswap pair addresses from using a RPC call to actually implementing the deterministic `CREATE2` address derivation in Rust. This saves us a dependent RPC call (the result is needed to create the `get_reserves` call). This reduces the amount of batches we need from 2 to 1.

### Test Plan
Unit test for create2 
E2E test to make sure we are still fetching the uniswap liquidity sources correctly.