Use different internal fee validity to reduce DB writes (gnosis/gp-v2-services#333)
Currently, we write one record per fee request to storage. @vkgnosis pointed out that this could become a DOS vector. As a countermeasure, this PR introduce the concept of an "internal validity" (two minutes) which is used to store records in the database, and an "external validity" of 1 minute, which is the official value we propagate to the client.

If a fee request for a token comes in we first check if we have already made an estimate for which the internal validity is still larger than the official validity we plan to report (that is if we have done an estimate for the same token in the last minute). In this case we reuse this estimate instead of creating and persisting a new one.

### Test Plan
Unit tests still work