Specify gas limit with a buffer (gnosis/gp-v2-services#359)
Fixes #354

We are seeing in practice that the gas estimate can be too tight. Therefore we should perform the estimate ourselves and multiply it with a factor to allow for a buffer (e.g. in case some state changes making the execution slightly more gas heavy at runtime).

This has the nice side effect of fixing a TODO for the gas estimation.

### Test Plan
Unit tests, observe no more out of gas failures.