Allow for Uniswap Spot price estimation (gnosis/gp-v2-services#360)
In #348 I handled the case when amount is 0 by just rounding it up to 1 (the smallest amount to buy/sell). While this leads to arithmetically valid results they are not very useful in practice because of integer division rounding errors for such small amounts (e.g if token B is worth more than token A, selling 1 of token A gives you 0 token B).

This PR adds functionality to the Pool struct, the uniswap shared methods and the price estimator to compute the spot price (relation of reserves) for a given path. This allows more accurate price estimation for 0 amounts which might also be useful when estimating the current value of a certain token (e.g. for display in the block explorer)

### Test Plan
Added unit tests for this case.