Move order signing logic into a trait for reuse (gnosis/gp-v2-services#384)
Because we plan to sign `OrderCancellation`, it will be important to refactor the order signing logic in a way that generalizes the common methods into a trait (`EIP712Signing`). This PR performs the refactor for `OrderCreation`so that less work needs to be done when introducing Order Cancellation.

I am not entirely sure it is appropriate to include this new unit test since it requires an ethcontract import into models. Perhaps now that the bug has been uncovered, we should really test `OrderBuilder.sign_with`