Cached Pool Fetcher (gnosis/gp-v2-services#376)
Now that we are using the uniswap pools in a lot of places (price estimate, fee estimate, base line solver, objective value computation), it might make sense to cache pool information as least until a new block arrives.

This PR introduces a CachedPoolFetcher which does exactly that. It uses the current_block_stream to invalidate its cache and otherwise performs a read_through logic with its inner fetcher.

I had to change current_block_stream to take a concrete Transport instead of a DynTransport. If there is value in having DynTransport we should probably change it everywhere in the codebase where we are passing Web3.

### Test Plan
Unit tests for the new component.