Suppress no valid path found alert (gnosis/gp-v2-services#406)
This PR suppresses the 'No valid path found alert', which we have seen from time to time last week in our alert channel. This happens if someone tries to trade tokens for which no baseline route to the fee native token exists.

We are already catching the case when no baseline between the two tokens exist (and map the Err into a None). However, e.g. when using very illiquid or non existent tokens there is also no path to the native token, we propagate up an error leading to a 500 response.

This PR makes it so that we also catch this case and turn the Err into a None, leading to a 404 resut.

### Test Plan
Local instance against xDAI:

http://localhost:8080/api/v1/fee?sellToken=0x1a5f9352af8af974bfc03399e3767df6370d82e4&buyToken=0x6810e776880c02933d47db1b9fc05908e5386b96&amount=100000000000000000000&kind=buy

> {"errorType":"NotFound","description":"Fee path for given Token pair was not found"}