Filter undercollateralized pools (gnosis/gp-v2-services#419)
The issue is explained in the comment and in [this thread](https://gnosisinc.slack.com/archives/CPZA1AGKY/p1616601928013300?thread_ts=1616597126.012000&cid=CPZA1AGKY)

Basically the actual token balance of a pool and the reserves stored in a pool can get out of sync for some weird but still popular enough tokens (e.g. AMPL). This PR makes it that we ignore pools that are out of sync.

In the real world, I assume the AMPL team manually fixes the reserves of the most liquid pool (AMPL/ETH) but may not care about very low liqiuidity pools which atm seem to not even work on native Uniswap (AMPL/USDC, AMPL/DAI when multihops disable in advanced settings).

![image](https://user-images.githubusercontent.com/1200333/112616773-69354e00-8e24-11eb-8edd-c0d290f8cd4f.png)


### Test Plan
e2e test is still passing...