Refactor settlement generation in solvers (gnosis/gp-v2-services#479)
This unfortunately large PR refactors how settlements are built in the solver implementations. Specifically, there is an inversion of control, where liquidity doesn't return things (trade and interactions) to be added to the settlement, but instead gets passed a `SettlementEncoder` and adds things directly to it. This is nice as:
- We can have a single `SettlementHandling` trait used for all liquidity
- Extending the settlement intermediate representation will not require changes to existing liquidity
- Solvers are now completely agnostic to how liquidity encodes itself in the solution, both `LimitOrder`s and `AmmOrder`s get added with the same `with_liquidity` method.

Furthermore, this opens the door for implementing a "settlement optimizer" which would take the intermediate `SettlementEncoder` representation and spit out an optimized solution (by, for example, collapsing interactions that can be collapsed).

This is based off of a discussion with @fleupold where we decided that having a "settlement builder" type interface would be beneficial and allow keeping solvers completely how liquidity actually get included into the settlement, only caring about how the liquidity is modeled.

Suggestions open to suggestions on how to make this smaller. Unfortunately, the `Settlement` struct and liquidity gets encoded into it was used in a lot of source and tests.

### Test Plan

CI still passes. This PR is a refactor, which modifies existing traits and how they are called.